create workaround for pallet_preimage version
Preimage pallet was added when we used Polkadot SDK v0.9.37, but
the migration provided by the pallet_preimage hasn't been applied.

However, the migration provided by the pallet_preimage is unnecessary
since we used the actual v1 types from the beginning. We only need
to align the reported storage version with the actual format
on the chain (already v1).